added logo

+++ public/javascripts/lists.js
@@ -36,8 +36,64 @@ window.onclick = function (event) {
 //----------------------------------Helper Functions To Fetch Calls To The Backend----------------------------------------
 
 //fetch all the lists from the server
-async function fetchLists() {
+// async function fetchLists() {
 
+//     const divLists = document.querySelector('#lists');
+//     try {
+//         const res = await fetch('/lists');
+//         if (!res.ok) {
+//             throw res;
+//         }
+//         const { lists, csrfToken } = await res.json();
+//         //create an anchor, rename btn, and delete btn to each list item
+//         lists.forEach(list => {
+
+//             const anchor = document.createElement('anchor');
+//             const btnRename = document.createElement('button');
+//             const btnDelete = document.createElement('button');
+//             const divContainer = document.createElement('div');
+
+//             divContainer.setAttribute('id', `div-container-${list.id}`);
+
+//             anchor.setAttribute('id', list.id);
+//             anchor.setAttribute('class', "list-anchors")
+//             anchor.setAttribute('href', `http://localhost:8080/lists/${list.id}`);
+//             btnRename.setAttribute('class', `list__btn--rename ${list.id}`);
+//             btnDelete.setAttribute('class', `list__btn--delete ${list.id}`);
+
+//             anchor.innerText = list.name;
+//             btnDelete.innerText = "Delete";
+
+//             //Create the modal for the rename btn
+//             btnRename.innerHTML = `Rename <div id="myModal-rename-${list.id}" class="modal-rename ${list.id}">
+//                                         <div class="modal-content-rename">
+//                                             <form class="list-rename-form">
+//                                                 <span class="close-rename">X</span>
+//                                                 <p class="add-list-rename">Rename List</p>
+//                                                 <label class="list-label-rename"> List name </label>
+//                                                 <br>
+//                                                 <input type="text" name="name" class="list-name-rename--input" required>
+//                                                 <br>
+//                                                 <input type="submit" value="Save" class="list-rename">
+//                                                 <input type="button" value="Cancel" id="cancel-${list.id}" class="list-cancel-rename">
+//                                         </div>
+//                                     </div>`;
+
+//             divLists.appendChild(divContainer);
+//             divContainer.appendChild(anchor);
+//             divContainer.appendChild(btnRename);
+//             divContainer.appendChild(btnDelete);
+//         });
+// 
+//         const inputCsurf = document.querySelector('#csurf')
+//         inputCsurf.setAttribute('value', csrfToken);
+//     } catch (error) {
+//         console.log(error);
+//     }
+// }
+
+
+async function fetchLists() {
     const divLists = document.querySelector('#lists');
     try {
         const res = await fetch('/lists');
@@ -47,160 +103,57 @@ async function fetchLists() {
         const { lists, csrfToken } = await res.json();
         //create an anchor, rename btn, and delete btn to each list item
         lists.forEach(list => {
-
-            const anchor = document.createElement('anchor');
-            const btnRename = document.createElement('button');
-            const btnDelete = document.createElement('button');
-            const divContainer = document.createElement('div');
-
-            divContainer.setAttribute('id', `div-container-${list.id}`);
-
-            anchor.setAttribute('id', list.id);
-            anchor.setAttribute('class', "list-anchors")
-            anchor.setAttribute('href', `http://localhost:8080/lists/${list.id}`);
-            btnRename.setAttribute('class', `list__btn--rename ${list.id}`);
-            btnDelete.setAttribute('class', `list__btn--delete ${list.id}`);
-
-            anchor.innerText = list.name;
-            btnDelete.innerText = "Delete";
-
-            //Create the modal for the rename btn
-            btnRename.innerHTML = `Rename <div id="myModal-rename-${list.id}" class="modal-rename ${list.id}">
-                                        <div class="modal-content-rename">
-                                            <form class="list-rename-form">
-                                                <span class="close-rename">X</span>
-                                                <p class="add-list-rename">Rename List</p>
-                                                <label class="list-label-rename"> List name </label>
-                                                <br>
-                                                <input type="text" name="name" class="list-name-rename--input" required>
-                                                <br>
-                                                <input type="submit" value="Save" class="list-rename">
-                                                <input type="button" value="Cancel" id="cancel-${list.id}" class="list-cancel-rename">
-                                        </div>
-                                    </div>`;
-
-            divLists.appendChild(divContainer);
-            divContainer.appendChild(anchor);
-            divContainer.appendChild(btnRename);
-            divContainer.appendChild(btnDelete);
+            if(list.name !== "_hidden") {
+                const anchor = document.createElement('anchor');
+                const btnRename = document.createElement('button');
+                const btnDelete = document.createElement('button');
+                const divContainer = document.createElement('div');
+                divContainer.setAttribute('id', `div-container-${list.id}`);
+                anchor.setAttribute('id', list.id);
+                anchor.setAttribute('class', "list-anchors")
+                anchor.setAttribute('href', `http://localhost:8080/lists/${list.id}`);
+                btnRename.setAttribute('class', `list__btn--rename ${list.id}`);
+                btnDelete.setAttribute('class', `list__btn--delete ${list.id}`);
+                anchor.innerText = list.name;
+                btnDelete.innerText = `Delete`;
+                //Create the modal for the rename btn
+                btnRename.innerHTML = `Rename <div id="myModal-rename-${list.id}" class="modal-rename ${list.id}">
+                                            <div class="modal-content-rename">
+                                                <form class="list-rename-form">
+                                                    <span class="close-rename">X</span>
+                                                    <p class="add-list-rename">Rename List</p>
+                                                    <label class="list-label-rename"> List name </label>
+                                                    <br>
+                                                    <input type="text" name="name" class="list-name-rename--input-${list.id}" required>
+                                                    <br>
+                                                    <input type="submit" value="Save" class="list-rename ${list.id}">
+                                                    <input type="button" value="Cancel" id="cancel-${list.id}" class="list-cancel-rename">
+                                            </div>
+                                        </div>`;
+                divLists.appendChild(divContainer);
+                divContainer.appendChild(anchor);
+                divContainer.appendChild(btnRename);
+                divContainer.appendChild(btnDelete);
+            };
         });
 
-        const inputCsurf = document.querySelector('#csurf')
-        inputCsurf.setAttribute('value', csrfToken);
-    } catch (error) {
-        console.log(error);
-    }
-}
-
 //-------------------------------------------------------------------------------
-
-const tasksContainer = document.querySelector('#tasks-section__tasks-container');
-let selectedTaskIds = new Set();
 //Display all tasks from the targeted list
 //Display all summary of the targeted list
-// async function fetchOneList(id) {
+async function fetchOneList(id) {
 
-//     // const ulTasks = document.querySelector('#tasks');
-//     try {
-//         // const res = await fetch(`/lists/${id}`);
-//         // if (!res.ok) {
-//         //     throw res;
-//         // }
-//         // const { list, tasks } = await res.json();
-//         const res = await fetch(`/tasks?listIds=${[id]}`);
-//         if (!res.ok) {
-//             throw res;
-//         }
-//         const json = await res.json();
-//         console.log(json);
-//         // tasks.forEach(task => {
-//         //     const li = document.createElement('li');
-//         //     li.innerHTML = task.name;
-//         //     ulTasks.appendChild(li);
-//         // });
-//     }
-//     catch (err) {
-//         console.log(err);
-//     }
-// }
-
-
-const getTasks = async (listIds = []) => {
-    const res = await fetch(`/tasks?listIds=${[...listIds]}`, {
-        method: 'GET',
-        headers: {
-            "Content-Type": "application/json"
+    // const ulTasks = document.querySelector('#tasks');
+    try {
+        const res = await fetch(`/lists/${id}`);
+        if (!res.ok) {
+            throw res;
         }
-    });
-    if (!res.ok)
-        throw res;
-    const { tasks } = await res.json();
-    return tasks;
+        const { list, tasks } = await res.json();
+    }
+    catch (err) {
+        console.log(err);
+    }
 }
-
-const filterTasks = async (tasks, query) => {
-    return tasks.filter(task => {
-        for (let prop in query) {
-            // Include must store an object with props: term and from
-            if (prop === 'include') {
-                const { term, from } = query[prop];
-                if (!task[from].toLowerCase().includes(term.toLowerCase()))
-                    return false;
-            }
-            // Exclude must store an object with props: term and from
-            else if (prop === 'exclude') {
-                const { term, from } = query[prop];
-                if (task[from].toLowerCase().includes(term.toLowerCase()))
-                    return false;
-            }
-            // Check if a query prop/value matches a task prop/value
-            else if (task[prop] !== query[prop])
-                return false;
-        }
-        return true;
-    });
-}
-
-const displayTasks = async (tasks) => {
-    if (!tasks) tasks = await getTasks();
-
-    // Reset selectedTaskIds
-    selectedTaskIds = new Set();
-
-    const tasksHtml = tasks.map(task => {
-        let taskStr = '';
-
-        if (task.sets) taskStr += `${task.sets} `;
-        if (task.reps) taskStr += `x ${task.reps} `;
-        taskStr += task.name;
-        if (task.duration) {
-            const dur = convertSeconds(task.duration);
-            taskStr += ` for ${dur[0]}:${dur[1]}:${dur[2]}`
-        }
-
-        return `<div id=task-${task.id} class="tasks-section__task">
-                    <div class="handle">
-                        <i class="fas fa-ellipsis-v"></i>
-                    </div>
-                    <input type="checkbox">
-                    <div class="card-text">${taskStr}</div>
-                </div>`;
-    });
-
-    tasksContainer.innerHTML = tasksHtml.join('');
-};
-
-const updateTasksSection = async (listIds, queries = []) => {
-    let tasks = await getTasks(listIds);
-    for (let query of queries)
-        tasks = await filterTasks(tasks, query);
-    await displayTasks(tasks);
-}
-
-
-
-
-
 
 //---------------------------------------------------------------------
 
@@ -364,10 +317,10 @@ window.addEventListener('DOMContentLoaded', async (event) => {
     const inputAddListBtn = document.querySelector('.list-add');
     const inputListValue = document.querySelector('.list-name--input');
 
-    const inputRenameSingle = document.querySelector('.list-rename');
-    const inputRenameValue = document.querySelector('.list-name-rename--input');
+    // const inputRenameSingle = document.querySelector('.list-rename');
+    // const inputRenameValue = document.querySelector('.list-name-rename--input');
 
-    const deleteBtns = document.querySelectorAll('.list__btn--delete');
+    // const deleteBtns = document.querySelectorAll('.list__btn--delete');
 
     //Listen for a click on each anchor element to display tasks
     listAnchors.forEach(list => {
@@ -375,7 +328,6 @@ window.addEventListener('DOMContentLoaded', async (event) => {
             event.preventDefault();
             event.stopPropagation();
             await updateTasksSection(event.target.id);
-            // await fetchOneList(event.target.id);
         })
     })
 
@@ -390,21 +342,37 @@ window.addEventListener('DOMContentLoaded', async (event) => {
     // listen for a click on the save btn to rename the list
     /*There is a bug where the modal will not pop up anymore when the
     rename btn is clicked only after creating a new list. */
-    inputRenameSingle.addEventListener('click', async (event) => {
-        event.stopPropagation();
-        event.preventDefault();
 
-        await updateList(renameBtnId, inputRenameValue.value);
-        modal_rename.style.display = "none";
-        inputRenameValue.value = "";
-    });
+    // inputRenameSingle.addEventListener('click', async (event) => {
+    //     event.stopPropagation();
+    //     event.preventDefault();
 
-    //Listen for a click on the delete btn
-    deleteBtns.forEach( btn => {
-        btn.addEventListener('click', async (event) => {
-            const deleteId = event.target.classList[1];
-            await deleteList(deleteId);
-        })
-    });
+    //     await updateList(renameBtnId, inputRenameValue.value);
+    //     modal_rename.style.display = "none";
+    //     inputRenameValue.value = "";
+    // });
+
+    // Listen for a click on the delete btn
+    // deleteBtns.forEach( btn => {
+    //     btn.addEventListener('click', async (event) => {
+    //         const deleteId = event.target.classList[1];
+    //         await deleteList(deleteId);
+    //     })
+    // });
+
+    //----------------------------------------------------------------------------------
+
+        const rootDiv = document.getElementById("lists")
+
+        rootDiv.addEventListener('click', async (e) => {
+            e.preventDefault()
+            let btn = e.target
+            console.log(btn)
+            let modalId
+            if(btn.classList.contains('list__btn--delete')){
+                let deleteId = e.target.classList[1];
+                await deleteList(deleteId);
+            }
+        }, true)
 
 });
